test(http): Extract core tests from `StatelessApplicationTest` to `StatelessApplicationCoreTest` class.
